tethering: User defined wifi/usb tether network. [3/3]
Enables users to select from a list of predefined networks
for use as the tether network, overriding the built-in networks.
This is useful, for example, where OpenVPN is in use with
routed setup.

Note: This commit includes a squashed version of
http://review.pac-rom.com/#/c/802/ pursuant to the permission of
the author (Lokesh Chamane <lokesh.c703@gmail.com>).

Change-Id: I0dce2379a3f85fc63e8140e0e8c9c4c8889c5ae2

Conflicts:
	res/values/pac_arrays.xml
	res/values/pac_strings.xml

Conflicts:
	res/values/cr_arrays.xml
	res/values/cr_strings.xml

Conflicts:
	res/values/cr_arrays.xml
	res/values/cr_strings.xml